# swagger.json
{"openapi":"3.0.0","info":{"title":"Na Minha Geladeira","version":"1.0.0","description":"API para gerenciar receitas, ingredientes e avaliações"},"paths":{"/usuarios/email/{email}":{"get":{"tags":["Usuarios"],"summary":"Busca um usuário por email","parameters":[{"name":"email","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"Usuário encontrado","content":{"application/json":{"schema":{"type":"object","properties":{"id":{"type":"integer"},"email":{"type":"string"},"hashSenha":{"type":"string"},"apelido":{"type":"string"},"adm":{"type":"boolean"}}}}}},"500":{"description":"Erro ao buscar usuário por email"}}}},"/usuarios/{id}":{"get":{"tags":["Usuarios"],"summary":"Busca um usuário por ID","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer"}}],"responses":{"200":{"description":"Usuário encontrado","content":{"application/json":{"schema":{"type":"object","properties":{"id":{"type":"integer"},"email":{"type":"string"},"hashSenha":{"type":"string"},"apelido":{"type":"string"},"adm":{"type":"boolean"}}}}}},"500":{"description":"Erro ao buscar usuário"}}},"put":{"tags":["Usuarios"],"summary":"Atualiza um usuário","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"type":"object","properties":{"hashSenha":{"type":"string"},"apelido":{"type":"string"}}}}}},"responses":{"200":{"description":"Usuário atualizado com sucesso","content":{"application/json":{"schema":{"type":"object","properties":{"id":{"type":"integer"},"email":{"type":"string"},"hashSenha":{"type":"string"},"apelido":{"type":"string"},"adm":{"type":"boolean"}}}}}},"500":{"description":"Erro ao atualizar usuário"}}}},"/receitas-por-ingredientes":{"get":{"tags":["Receitas"],"summary":"Busca receitas com base em uma lista de ingredientes","parameters":[{"name":"ingredientes","in":"query","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"Lista de receitas","content":{"application/json":{"schema":{"type":"array","items":{"type":"object","properties":{"id":{"type":"integer"},"nome":{"type":"string"},"descricao":{"type":"string"},"idDoUsuario":{"type":"integer"},"IngredienteReceita":{"type":"array","items":{"type":"object","properties":{"nomeDoIngrediente":{"type":"string"},"quantidade":{"type":"number"}}}}}}}}}},"500":{"description":"Erro ao buscar receitas por ingredientes"}}}},"/ingredientes":{"get":{"tags":["Ingredientes"],"summary":"Busca todos os ingredientes","responses":{"200":{"description":"Lista de ingredientes","content":{"application/json":{"schema":{"type":"array","items":{"type":"object","properties":{"nome":{"type":"string"},"tipoDeMedida":{"type":"string","enum":["UNIDADES","GRAMAS","MILILITROS"]}}}}}}},"500":{"description":"Erro ao buscar ingredientes"}}},"post":{"tags":["Ingredientes"],"summary":"Cria um novo ingrediente","requestBody":{"required":true,"content":{"application/json":{"schema":{"type":"object","properties":{"nome":{"type":"string"},"tipoDeMedida":{"type":"string","enum":["UNIDADES","GRAMAS","MILILITROS"]}}}}}},"responses":{"201":{"description":"Ingrediente criado com sucesso","content":{"application/json":{"schema":{"type":"object","properties":{"nome":{"type":"string"},"tipoDeMedida":{"type":"string","enum":["UNIDADES","GRAMAS","MILILITROS"]}}}}}},"500":{"description":"Erro ao criar ingrediente"}}}},"/ingredientes/nome/{nome}":{"get":{"tags":["Ingredientes"],"summary":"Busca um ingrediente pelo nome","parameters":[{"name":"nome","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"Ingrediente encontrado","content":{"application/json":{"schema":{"type":"object","properties":{"nome":{"type":"string"},"tipoDeMedida":{"type":"string","enum":["UNIDADES","GRAMAS","MILILITROS"]}}}}}},"500":{"description":"Erro ao buscar ingrediente pelo nome"}}}},"/avaliacoes/{idDaReceita}":{"get":{"tags":["Avaliações"],"summary":"Busca avaliações por ID da receita","parameters":[{"name":"idDaReceita","in":"path","required":true,"schema":{"type":"integer"}}],"responses":{"200":{"description":"Lista de avaliações","content":{"application/json":{"schema":{"type":"array","items":{"type":"object","properties":{"idDoUsuario":{"type":"integer"},"idDaReceita":{"type":"integer"},"nota":{"type":"number"},"comentario":{"type":"string"}}}}}}},"500":{"description":"Erro ao buscar avaliações"}}}},"/avaliacoes":{"post":{"tags":["Avaliações"],"summary":"Cria uma nova avaliação","requestBody":{"required":true,"content":{"application/json":{"schema":{"type":"object","properties":{"idDoUsuario":{"type":"integer"},"idDaReceita":{"type":"integer"},"nota":{"type":"number"},"comentario":{"type":"string"}}}}}},"responses":{"201":{"description":"Avaliação criada com sucesso","content":{"application/json":{"schema":{"type":"object","properties":{"idDoUsuario":{"type":"integer"},"idDaReceita":{"type":"integer"},"nota":{"type":"number"},"comentario":{"type":"string"}}}}}},"500":{"description":"Erro ao criar avaliação"}}}},"/avaliacoes/{idDoUsuario}/{idDaReceita}":{"put":{"tags":["Avaliações"],"summary":"Atualiza uma avaliação","parameters":[{"name":"idDoUsuario","in":"path","required":true,"schema":{"type":"integer"}},{"name":"idDaReceita","in":"path","required":true,"schema":{"type":"integer"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"type":"object","properties":{"nota":{"type":"number"},"comentario":{"type":"string"}}}}}},"responses":{"200":{"description":"Avaliação atualizada com sucesso","content":{"application/json":{"schema":{"type":"object","properties":{"idDoUsuario":{"type":"integer"},"idDaReceita":{"type":"integer"},"nota":{"type":"number"},"comentario":{"type":"string"}}}}}},"500":{"description":"Erro ao atualizar avaliação"}}},"delete":{"tags":["Avaliações"],"summary":"Deleta uma avaliação","parameters":[{"name":"idDoUsuario","in":"path","required":true,"schema":{"type":"integer"}},{"name":"idDaReceita","in":"path","required":true,"schema":{"type":"integer"}}],"responses":{"204":{"description":"Avaliação deletada com sucesso"},"500":{"description":"Erro ao deletar avaliação"}}}}},"components":{},"tags":[{"name":"Usuarios","description":"Gerencia usuários."},{"name":"Ingredientes","description":"Gerencia ingredientes."},{"name":"Avaliações","description":"Gerencia avaliações."}]}